scripts: Create target-runnable binaries
These binaries will allow for on-device compilation of kernel modules. This
includes versioning of kernel modules for commonly-used binary blob wrapping
techniques.

Note that these are simply changes to generate the on-device runnable binaries,
it won't suddenly make things like dkms work on your ARM board.

The resulting binaries will be prefixed with the same prefix as your
cross-compile toolchain.

Signed-off-by: Brad Mouring <brad.mouring@ni.com>
Acked-by: Joe Hershberger <joseph.hershberger@ni.com>
Acked-by: Irwan Djajadi <irwan.djajadi@ni.com>
Acked-by: Will Page <will.page@ni.com>