igb: don't drop packets on VLAN 0 when VLAN is not configured
A VLAN ID of 0 is intended to convey only user priority information, and
packets with this ID should be received even if VLAN is not configured.
This is necessary in order to support industrial communications
protocols such as IEC 61850 (GOOSE).

This patch is based on a patch from Alexander Duyck on the thread at
http://permalink.gmane.org/gmane.linux.drivers.e1000.devel/8624

Signed-off-by: Ben Shelton <ben.shelton@ni.com>