Marvell phy: read LED settings from device tree
The m88e1121_config_aneg function is extended to check if the "leds"
register is defined in the phy section of the device tree, and if that
register contains a valid data that value is set in the LED[2:0]
Function Control Register. If the "leds" register is not defined, the
default value is written to the LED[2:0] Function Control Register.

Signed-off-by: Mihaly Varga <mihaly.varga@ni.com>
Acked-by: Ben Shelton <ben.shelton@ni.com>
Acked-by: Jaeden Amero <jaeden.amero@ni.com>
Natinst-ReviewBoard-ID: 81628