sdhci-of-arasan: Follow mmc bindings
The documentation for the OF bindings for sdhci-of-arasan
(Documentation/devicetree/bindings/mmc/arasan,sdhci.txt) stated that the
bindings follow the mmc bindings. However, mmc_of_parse was never called
and thus those bindings were never actually followed.

It is useful to be able to follow the mmc bindings, so let's do that.

In following the mmc bindings, we have to supply our own sdhci op
"get_max_clock", otherwise, the sdhci driver will overwrite the
mmc_of_parse-supplied value for 'max-frequency'. It is desirable to
specify a 'max-frequency' (for digital signal integrity reasons), so
we'll use the 'max-frequency' value if non-zero, and the maximum of the
underlying clock otherwise.

Signed-off-by: Jaeden Amero <jaeden.amero@ni.com>
Acked-by: Ben Shelton <ben.shelton@ni.com>
Acked-by: Brad Mouring <brad.mouring@ni.com>
Acked-by: Gratian Crisan <gratian.crisan@ni.com>
Natinst-ReviewBoard-ID: 84053