HACK NI: devicetree: Use amba@0 instead of amba
The device tree path to the USB device changed when "amba@0" changed to
"amba". This change reverts back to "amba@0" so that bootloaders which
assumed the USB device tree path was stable will still work.

The correct solution, and why this is a hack, is to make the bootscript
take over the responsibility, from U-Boot, of switching the USB device
between host and device mode. The responsibility should have been with
the bootscript and not in U-Boot from the first implementation of this
feature. Without some sort of fix (e.g. this amba renaming hack), USB
can't switch between host and device mode.

Since U-Boot can't anticipate the itb package contents and layouts, the
bootscript must set the USB host or device mode for controllers that can
dynamically switch USB-mode (between host and device) at boot time. It
is incorrect for U-Boot to attempt to do this on its own, as it is
destined to not always do the right thing since the device tree
interface isn't stable.

Example of U-Boot getting it wrong: Some controllers have a U-Boot that
attempts to set dr_mode on "/amba@0/usb@e0002000" (which works for
pre-LabVIEW RT 15.0), but device tree changed at one point in time such
that the correct path was "/amba/usb@e0002000" (LabVIEW RT 15.0).

So, a more proper solution that this hack is to do the following:
1. Figure out which USB device U-Boot wanted to set dr_mode on
(CONFIG_OTG_USB_DT_NODE).
2. Modify dr_mode on behalf of U-Boot in the bootscript (like we should
have done in the first place, as the bootscript will always understand
the device tree paths).

Future authors of a more proper solution are lucky that our shipped
U-boot scripts will gracefully fail when they attempt to use the wrong
device tree path. So, their solution doesn't have to do any cleanup or
fixing on the bootloader's behalf in the bootscript. They simply have to
figure out the right thing to do, and then do it.

See <http://review-board.natinst.com/r/93029/> for an attempt at a more
proper solution. Note that because this HACK is cleaner, we went with
this solution for now. We may have to change at a later time to
something more complicated and gross.

Signed-off-by: Jaeden Amero <jaeden.amero@ni.com>
Reviewed-by: Josh Cartwright <joshc@ni.com>
Natinst-ReviewBoard-ID: 95820