net/phy: micrel: Add workaround for autoneg failure
In addition to the issue in 'net/phy: micrel: Add workaround for bad
autoneg', sometimes the micrel KSZ9031 can fail to negotiate at all.
When this happens, the phy will show link partner advertisements but
will not show the link coming up.

To fix this, reset the PHY if it has been down with link partner
advertisements for more than 20 seconds.

Signed-off-by: Nathan Sullivan <nathan.sullivan@ni.com>
Signed-off-by: Brad Mouring <brad.mouring@ni.com>
Acked-by: Josh Cartwright <josh.cartwright@ni.com>
Acked-by: Tony Liechty <tony.liechty@ni.com>
Natinst-ReviewBoard-ID: 118397